Extract  Class pattern: Move MerkleTree (from MerkleTransaction) to its own class file inside core.crypto package.
